update jenkins main

+++ Jenkinsfile
@@ -33,19 +33,23 @@ pipeline {
         DOCKER_TAG="${GIT_BRANCH.tokenize('/').pop()}-${GIT_COMMIT.substring(0,7)}"
       }
       steps {
-        sh "echo build php - laravel"
-        sh "docker build -t ${DOCKER_IMAGE_PHP}:${DOCKER_TAG} -f Dockerfile.php . "
-        sh "docker tag ${DOCKER_IMAGE_PHP}:${DOCKER_TAG} ${DOCKER_IMAGE_PHP}:latest"
-        sh "docker image ls | grep ${DOCKER_IMAGE_PHP}"
-        withCredentials([usernamePassword(credentialsId: 'dockerhub-token', usernameVariable: 'DOCKER_USERNAME', passwordVariable: 'DOCKER_PASSWORD')]) {
-            sh 'echo $DOCKER_PASSWORD | docker login --username $DOCKER_USERNAME --password-stdin'
-            sh "docker push ${DOCKER_IMAGE_PHP}:${DOCKER_TAG}"
-            sh "docker push ${DOCKER_IMAGE_PHP}:latest"
-        }
+        script {
+          if (GIT_BRANCH == 'main') {
+            sh "echo build php - laravel"
+            sh "docker build -t ${DOCKER_IMAGE_PHP}:${DOCKER_TAG} -f Dockerfile.php . "
+            sh "docker tag ${DOCKER_IMAGE_PHP}:${DOCKER_TAG} ${DOCKER_IMAGE_PHP}:latest"
+            sh "docker image ls | grep ${DOCKER_IMAGE_PHP}"
+            withCredentials([usernamePassword(credentialsId: 'dockerhub-token', usernameVariable: 'DOCKER_USERNAME', passwordVariable: 'DOCKER_PASSWORD')]) {
+                sh 'echo $DOCKER_PASSWORD | docker login --username $DOCKER_USERNAME --password-stdin'
+                sh "docker push ${DOCKER_IMAGE_PHP}:${DOCKER_TAG}"
+                sh "docker push ${DOCKER_IMAGE_PHP}:latest"
+            }
 
-        //clean to save disk
-        sh "docker image rm ${DOCKER_IMAGE_PHP}:${DOCKER_TAG}"
-        sh "docker image rm ${DOCKER_IMAGE_PHP}:latest"
+            //clean to save disk
+            sh "docker image rm ${DOCKER_IMAGE_PHP}:${DOCKER_TAG}"
+            sh "docker image rm ${DOCKER_IMAGE_PHP}:latest"
+          }
+        }
       }
     }
